requirements.txt
aiohttp==3.6.3
appdirs==1.4.4
APScheduler==3.6.3
async-timeout==3.0.1
attrs==20.3.0
better-profanity==0.7.0
certifi==2020.6.20
chardet==3.0.4
conda==4.3.16
discord-ext-menus==1.0.0a27+g6f2b873
discord.py==1.5.1
distlib==0.3.1
filelock==3.0.12
idna==2.10
multidict==4.7.6
pipenv==2020.8.13
pycosat==0.6.3
pytz==2020.4
requests==2.24.0
ruamel.yaml==0.16.12
ruamel.yaml.clib==0.2.2
six==1.15.0
tzlocal==2.1
urllib3==1.25.10
virtualenv==20.0.31
virtualenv-clone==0.5.4
yarl==1.5.1
